chore: workaround for broken CircleCI badges
CircleCI and Shields.io report a build as "broken" if it was skipped. Unfortunately for users of Semantic Release, this means that any substantial commit to `master` makes their build look like it's broken 😱

Solution: Run a (no-op) build on the `chore` commit following a release, just to get the green badge. 😕